app/backend/requirements.txt
agno
streamlit
fastapi
flask-cors
google-generativeai
yfinance
duckduckgo-search
newspaper4k
plotly
pandas
numpy
reportlab
jinja2
python-dotenv
requests
beautifulsoup4
google-genai
gunicorn